remove fake repository and move stub services to root
