+++ order/command/pack_item_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/segmentio/kafka-go"
 	"github.com/stretchr/testify/assert"
 	i "mc-burger-orders/item"
+	"mc-burger-orders/order"
 	m "mc-burger-orders/order/model"
 	"mc-burger-orders/stack"
 	"testing"
@@ -52,8 +53,8 @@ func shouldPackItemPointedInMessage(t *testing.T) {
 
 	message := givenKafkaMessage(t, messageValue)
 
-	kitchenService := &StubService{}
-	repositoryStub := &StubRepository{o: givenExistingOrder(), methodCalled: make([]map[string]any, 0)}
+	kitchenService := &order.StubService{}
+	repositoryStub := &order.StubRepository{o: givenExistingOrder(), methodCalled: make([]map[string]any, 0)}
 
 	sut := &PackItemCommand{
 		Stack:          s,
@@ -124,8 +125,8 @@ func shouldFinishPackingOrderWhenLastItemsCameFromKitchen(t *testing.T) {
 
 	message := givenKafkaMessage(t, messageValue)
 
-	kitchenService := &StubService{}
-	repositoryStub := &StubRepository{o: givenExistingOrder(), methodCalled: make([]map[string]any, 0)}
+	kitchenService := &order.StubService{}
+	repositoryStub := &order.StubRepository{o: givenExistingOrder(), methodCalled: make([]map[string]any, 0)}
 
 	sut := &PackItemCommand{
 		Stack:          s,
@@ -184,8 +185,8 @@ func shouldRequestAdditionalItemWhenMoreAreNeeded(t *testing.T) {
 
 	message := givenKafkaMessage(t, messageValue)
 
-	kitchenService := &StubService{}
-	repositoryStub := &StubRepository{o: givenExistingOrder(), methodCalled: make([]map[string]any, 0)}
+	kitchenService := &order.StubService{}
+	repositoryStub := &order.StubRepository{o: givenExistingOrder(), methodCalled: make([]map[string]any, 0)}
 
 	sut := &PackItemCommand{
 		Stack:          s,
@@ -231,8 +232,8 @@ func shouldFailWhenMessageValueIsEmpty(t *testing.T) {
 	s := stack.NewStack(stack.CleanStack())
 	message := givenKafkaMessage(t, make([]map[string]any, 0))
 
-	kitchenService := &StubService{}
-	repositoryStub := &StubRepository{o: givenExistingOrder()}
+	kitchenService := &order.StubService{}
+	repositoryStub := &order.StubRepository{o: givenExistingOrder()}
 
 	sut := &PackItemCommand{
 		Stack:          s,
@@ -270,8 +271,8 @@ func shouldFailWhenMessageKeyMissingOrderNumber(t *testing.T) {
 		Value: b,
 	}
 
-	kitchenService := &StubService{}
-	repositoryStub := &StubRepository{o: givenExistingOrder()}
+	kitchenService := &order.StubService{}
+	repositoryStub := &order.StubRepository{o: givenExistingOrder()}
 
 	sut := &PackItemCommand{
 		Stack:          s,
@@ -309,8 +310,8 @@ func shouldFailWhenMessageKeyHasInvalidOrderNumber(t *testing.T) {
 		Value: b,
 	}
 
-	kitchenService := &StubService{}
-	repositoryStub := &StubRepository{o: givenExistingOrder()}
+	kitchenService := &order.StubService{}
+	repositoryStub := &order.StubRepository{o: givenExistingOrder()}
 
 	sut := &PackItemCommand{
 		Stack:          s,
