+++ order/stub_service_test.go
@@ -1,4 +1,4 @@
-package command
+package order
 
 import (
 	"context"
